backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose[cryptography]
bcrypt>=4.0.0
pydantic
